Rework random generators
Now the randomPointGenerator is renamed regular as it generates a point at a regular distance from origin.
Another one is used for bezier surface generation, to handle the offset in the other dimension of the surface.
It creates a random point from an origin, but not regularly, fully random in every direction.